refactor: extract demo mode toggle out of auth_repository
The four service files imported isDemoProvider from auth_repository.dart
so their providers could pick mock vs. real implementations. But
auth_repository.dart imports portal_service.dart and student_query_service.dart
for the service types, which made services and the repository
mutually dependent.

Move DemoNotifier, isDemoProvider, demoUsername, and isDemoCredentials()
to lib/services/demo_mode.dart. Services and the repository now both
import this file; the import direction is one-way (services → demo_mode,
repository → demo_mode + services).
