+++ lib/services/student_query/student_query_service.dart
@@ -3,7 +3,7 @@ import 'package:tattoo/models/course.dart';
 import 'package:tattoo/models/ranking.dart';
 import 'package:tattoo/models/score.dart';
 import 'package:tattoo/models/user.dart';
-import 'package:tattoo/repositories/auth_repository.dart';
+import 'package:tattoo/services/demo_mode.dart';
 import 'package:tattoo/services/student_query/mock_student_query_service.dart';
 import 'package:tattoo/services/student_query/ntut_student_query_service.dart';
